remove class based code, create header and home components
