+++ src/components/Calculator.js
@@ -1,35 +1,22 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import Row from "./Row";
 import calculate from "../logic/calculate";
 import "../styles/calculator.css";
 
-class Calculator extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      dataObj: {
-        total: null,
-        next: null,
-        operation: null,
-      },
-    };
-    this.updateDataObj = this.updateDataObj.bind(this);
-  }
+export default function CalculatorHook() {
+  const [dataObj, setDataObj] = useState({
+    total: null,
+    next: null,
+    operation: null,
+  });
 
-  componentDidUpdate() {
-    this.updateScreen();
-  }
-
-  updateDataObj(event) {
+  function updateDataObj(event) {
     const value = event.target.textContent;
-    this.setState((prevState) => ({
-      dataObj: { ...calculate(prevState.dataObj, value) },
-    }));
+    setDataObj((prevDataObj) => calculate(prevDataObj, value));
   }
 
-  updateScreen() {
+  function updateScreen() {
     const screen = document.querySelector(".screen");
-    const { dataObj } = this.state;
     if (dataObj.next && !dataObj.total) {
       screen.textContent = dataObj.next;
     } else if (!dataObj.next && dataObj.total) {
@@ -49,49 +36,49 @@ class Calculator extends React.Component {
     }
   }
 
-  render() {
-    return (
-      <div id="calculator">
-        <div className="screen" />
-        <div className="buttons">
-          <Row
-            updateDataObj={this.updateDataObj}
-            first="AC"
-            second="+/-"
-            third="%"
-            fourth="÷"
-          />
-          <Row
-            updateDataObj={this.updateDataObj}
-            first="7"
-            second="8"
-            third="9"
-            fourth="x"
-          />
-          <Row
-            updateDataObj={this.updateDataObj}
-            first="4"
-            second="5"
-            third="6"
-            fourth="-"
-          />
-          <Row
-            updateDataObj={this.updateDataObj}
-            first="1"
-            second="2"
-            third="3"
-            fourth="+"
-          />
-          <Row
-            updateDataObj={this.updateDataObj}
-            first="0"
-            second="."
-            third="="
-          />
-        </div>
-      </div>
-    );
-  }
-}
+  useEffect(() => {
+    updateScreen();
+  }, [dataObj]);
 
-export default Calculator;
+  return (
+    <div id="calculator">
+      <div className="screen" />
+      <div className="buttons">
+        <Row
+          updateDataObj={updateDataObj}
+          first="AC"
+          second="+/-"
+          third="%"
+          fourth="÷"
+        />
+        <Row
+          updateDataObj={updateDataObj}
+          first="7"
+          second="8"
+          third="9"
+          fourth="x"
+        />
+        <Row
+          updateDataObj={updateDataObj}
+          first="4"
+          second="5"
+          third="6"
+          fourth="-"
+        />
+        <Row
+          updateDataObj={updateDataObj}
+          first="1"
+          second="2"
+          third="3"
+          fourth="+"
+        />
+        <Row
+          updateDataObj={updateDataObj}
+          first="0"
+          second="."
+          third="="
+        />
+      </div>
+    </div>
+  );
+}
