+++ src/components/Header.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Header = () => {
+  return (
+    <header>
+      <h1>Math Magicians</h1>
+      <ul>
+        <li>Home</li>
+        <li>Calculator</li>
+        <li>Quote</li>
+      </ul>
+    </header>
+  )
+}
+
+export default Header
